[Tests] Restructure Cell Callback Test
This test used to rely on a set of global “labels” and callbacks. The test would check which types with which label had invoked a callback.

This relied a lot on global state and made the test hard to read. This change introduces a mock cell that allows tracking calls that the cell received while not relying on global state.