Remove TableSectionViewModel.collapsed, close #120
This is seems to be a vestigial aspect of the framework from early PlanGrid days.

We don't use this anywhere in PlanGrid.

To achieve this sort of "collapsing" behavior, you'd capture this state elsewhere in your model, and simply regenerate your SectionModel with zero CellViewModels.

This is how we currently do this in PlanGrid.

And, I think this makes more sense to generate full state changes, rather than "hiding" a section. (i.e., this is sort of a hack)

Further, this is complicates notions of "empty/nil state" (see #48 and IP-3275) and I want to remove that complexity.

Closes #120