reloadData if diffing before refreshViews if diffing is disabled (#127)
This calls reloadData on self.tableView when diffing is disabled. This ensures that the UITableView is in sync with new data model before attempting to call methods on UITableView that would require the data model to be in sync to get correct results (i.e. cellForRow(at:)). This prevents crashing in cellForRow(at:) when UITableView asks for cells near a user-requested (programmatically), when the data model may not actually have the data for that cell.